feat(plus): song visibility behaviors for collaborative playlists

+++ app/Services/PlaylistCollaborationService.php
@@ -2,6 +2,12 @@
 
 namespace App\Services;
 
+use App\Events\NewPlaylistCollaboratorJoined;
+use App\Exceptions\CannotRemoveOwnerFromPlaylistException;
+use App\Exceptions\KoelPlusRequiredException;
+use App\Exceptions\NotAPlaylistCollaboratorException;
+use App\Exceptions\PlaylistCollaborationTokenExpiredException;
+use App\Exceptions\SmartPlaylistsAreNotCollaborativeException;
 use App\Facades\License;
 use App\Models\Playlist;
 use App\Models\PlaylistCollaborationToken;
@@ -9,14 +15,14 @@ use App\Models\User;
 use App\Values\PlaylistCollaborator;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\DB;
-use Webmozart\Assert\Assert;
 
 class PlaylistCollaborationService
 {
     public function createToken(Playlist $playlist): PlaylistCollaborationToken
     {
         self::assertKoelPlus();
-        Assert::false($playlist->is_smart, 'Smart playlists are not collaborative.');
+
+        throw_if($playlist->is_smart, SmartPlaylistsAreNotCollaborativeException::class);
 
         return $playlist->collaborationTokens()->create();
     }
@@ -28,13 +34,17 @@ class PlaylistCollaborationService
         /** @var PlaylistCollaborationToken $collaborationToken */
         $collaborationToken = PlaylistCollaborationToken::query()->where('token', $token)->firstOrFail();
 
-        Assert::false($collaborationToken->expired, 'The token has expired.');
+        throw_if($collaborationToken->expired, PlaylistCollaborationTokenExpiredException::class);
 
         if ($collaborationToken->playlist->ownedBy($user)) {
             return $collaborationToken->playlist;
         }
 
         $collaborationToken->playlist->addCollaborator($user);
+
+        // Now that we have at least one external collaborator, the songs in the playlist should be made public.
+        // Here we dispatch an event for that to happen.
+        event(new NewPlaylistCollaboratorJoined($user, $collaborationToken));
 
         return $collaborationToken->playlist;
     }
@@ -55,6 +65,9 @@ class PlaylistCollaborationService
     {
         self::assertKoelPlus();
 
+        throw_if($user->is($playlist->user), CannotRemoveOwnerFromPlaylistException::class);
+        throw_if(!$playlist->hasCollaborator($user), NotAPlaylistCollaboratorException::class);
+
         DB::transaction(static function () use ($playlist, $user): void {
             $playlist->collaborators()->detach($user);
             $playlist->songs()->wherePivot('user_id', $user->id)->detach();
@@ -63,6 +76,6 @@ class PlaylistCollaborationService
 
     private static function assertKoelPlus(): void
     {
-        Assert::true(License::isPlus(), 'Playlist collaboration is only available with Koel Plus.');
+        throw_unless(License::isPlus(), KoelPlusRequiredException::class);
     }
 }
